Add bioc-to-text pipeline
This is the initial commit of the bioc(xml)-to-plain-text pipeline. The pipeline consumes a directory of BioC XML files and outputs a plain text version of the document and a document containing section annotations (in BioNLP format). The BioC XML file, plain text file, and annotation file all get stored in Cloud Datastore.