implemented versioning of documents
Versioning of documents based on pipeline name and version has been implemented. The pom.xml version is now automatically added to the jar META-INF/MANIFEST.MF file, and it is this version that is used by the code. Note, the code must use the jar file for this to work, i.e. 'java -jar ...' is required.