Updated ExtractContentFn
Renamed from ExtractTextFn to signify that it can be used to load more than just text files. If target ProcessingStatusFlag is TEXT_DONE a new status entity is created, otherwise an existing status entity for the document is assumed to exist and is updated upon successful completion of the load.