Initial SAX parser implementation
This commit constitutes an initial implementation of the MedlineXmlToTextPipeline. It is inefficient with regard to reading the input XML files. The current implementation requires the files to be loaded into memory entirely (and this may be problematic as they are 30GB compressed). Further development will involve using the Beam XmlIO mechanism which should provide more appropriate handling of the input files.