Filter duplicate entities/documents/failures prior to Datastore load
ProcessingStatusToEntityFn has been updated to return a KV<String, Entity> instead of just the status entity. The String is the entity key, and allows for duplicates to be detected prior to Datastore insertion/update.

Pipelines that create the status entities (the pipelines that create/load plain text into Datastore) have been updated to filter out duplicates prior to Datastore insertion.

Why is this necessary?  DatastoreIO does non-transactional writing into Datastore. If the same entity is written more than once in the same commit, an error occurs: "A non-transactional commit may not contain multiple mutations affecting the same entity." To avoid this error, which results in a pipeline stalling, we filter out any duplicate entities prior to loading them into Datastore.  Files are supposed to be deterministic sources for Beam pipelines, so it is unclear at the moment why we are observing redundant status entities, but the fix in this commit seems to solve the issue.

Note, similar changes have also been made to the failure machinery to prevent issues with duplicates as well as instances where document content is being loaded into Datastore.