Streamline datastore processing status update
Streamlined the updating of the status entity in Datastore (to record when a job was completed for a particular document) by passing the entity through the pipeline. Previously only the document ID was passed between pipeline steps, causing the entity to be re-looked up at the completion of processing. This lookup and update is time intensive as it did not use the DatastoreIO.v1 API. The newly implemented solution should be significantly faster.