initial commit of ontology-to-kgx machinery
This code creates KGX files using OWL or OBO files as input. The initial thought was that these KGs could supplement our text-mined KGs by providing the ontology subsumption heirarchies and other specified relations. This was created in preparation for the September 2020 relay. At the relay it was announced that there will be an ontology KP, so this code may no longer be needed.