update BiocToTextConverter
Previously it was adding line breaks in order to get the section offsets to match the numbers provided in the BioC XML files. This resulted in text files that had excessive line breaks in some cases. The change made here substitutes spaces instead of line breaks. It's unclear if it's truly necessary to match the offsets reported in the BioC files but for now we will keep it that way.