Fixed bug in sentence identifier
The hash did not include document zone previously, however there are cases where a sentence appears multiple times in a document, but in different zones, e.g. title and abstract.

Also added validation to the ClassifiedSentenceStoragePipeline such that all values meet the database field criteria, e.g. make sure none of the String fields are too long to fit in the DB.