abbreviation detection fix
Original test for when an Ab3P output line is an abbreviation line was too simplistic as it was just counting the number of pipes on the line. This lead to NullPointerExceptions when processing some documents that had pipes on lines (usually in equations). The code that detects abbreviation lines should now be more robust.