new version of turku docker container
This new version builds off of a Turku docker container instead of trying to build from scratch. I still was not able to get the CRAFT model to work, but my hope is that this english model is sufficient, and robust. I have left my attempt at building a container for the CRAFT model from scratch as it works to a degree (it errors at times and stops processing prematurely). For now, the Dockerfile.en.worker is the one to use. (en = english)