augmented text revisions
instead of storing the document text redundantly - we store the augmented text separately and add it to the original document text when needed (oger). Similarly, we create an augmented sentences doc that is separate from the original sentences and stored separately - this is used as input to the crf component.